// src/payload-types.ts
export interface Image {
  id: string;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "portfolios".
 */
export interface Portfolio {
  id: string;
  name: string;
  image?: (string | null) | Image;
  links?: Link[] | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "Link".
 */
export interface Link {
  text: string;
  url: string;
  type: 'internal' | 'external' | 'email' | 'phone';
  id?: string | null;
  blockName?: string | null;
  blockType: 'link_block';
}
// ...
export interface GitHubSegmentBlock {
  title?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  image?: (string | null) | Image;
  call_to_action?: Link[] | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'github_segment';
}